feat: Add DataStory controls and related components
