+++ packages/ui/src/components/DataStory/controls/AddNodeControl.tsx
@@ -1,0 +1,17 @@
+import { ControlButton } from '@xyflow/react';
+import { AddNodeIcon } from '../icons/addNodeIcon';
+
+export const AddNodeControl = ({ setShowAddNode }: { setShowAddNode: (show: boolean) => void }) => (
+  <ControlButton
+    onClick={() => setShowAddNode(true)}
+    title="Add Node"
+    data-cy="add-node-button"
+    aria-label="addNode"
+  >
+    <AddNodeIcon/>
+  </ControlButton>
+);
+
+AddNodeControl.defaultProps = {
+  ariaLabel: 'addNode'
+};
