+++ packages/ui/src/components/DataStory/controls/SaveControl.tsx
@@ -1,0 +1,21 @@
+import { ControlButton } from '@xyflow/react';
+import { SaveIcon } from '../icons/saveIcon';
+import { useDataStoryControls } from './DataStoryControls';
+
+export const SaveControl = () => {
+  const { onSave, getDiagram } = useDataStoryControls();
+  
+  return (
+    <ControlButton
+      title="Save"
+      aria-label="save"
+      onClick={() => onSave?.(getDiagram())}
+    >
+      <SaveIcon/>
+    </ControlButton>
+  );
+};
+
+SaveControl.defaultProps = {
+  ariaLabel: 'save'
+};
