+++ packages/ui/src/index.ts
@@ -4,13 +4,13 @@ export { DataStoryCanvasProvider } from './components/DataStory/store/store';
 export { DropDown } from './components/DropDown';
 export { SaveIcon } from './components/DataStory/icons/saveIcon'
 export { OpenIcon } from './components/DataStory/icons/openIcon'
-export { useDataStoryControls } from './components/DataStory/controls/dataStoryControls';
+export { useDataStoryControls } from './components/DataStory/controls/DataStoryControls';
 export { eventManager, useDataStoryEvent } from './components/DataStory/events/eventManager'
 export { DataStoryEvents, type DataStoryEventType } from './components/DataStory/events/dataStoryEventType'
 export type { DataStoryProps, ClientRunParams, ServerClientObservationConfig } from './components/DataStory/types'
 export { default as NodeComponent } from './components/Node/NodeComponent';
 export { createJSClient } from './components/DataStory/clients/createJSClient';
-export type { DataStoryControlsType } from './components/DataStory/controls/dataStoryControls';
+export type { DataStoryControlsType } from './components/DataStory/controls/DataStoryControls';
 export type { WorkspaceApiClientImplement } from './components/DataStory/clients/WorkspaceApiClientImplement';
 export { processWaitingResponse, waitForResponse } from './components/DataStory/clients/WebSocketHandleResponseMiddleware';
 export { createSocketClient } from './components/DataStory/clients/createSocketClient';
